Added game logic functions
