+++ .vscode/settings.json
@@ -1,5 +1,6 @@
 {
     "files.associations": {
-        "board.h": "c"
+        "board.h": "c",
+        "instance.h": "c"
     }
 }
