+++ end.h
@@ -1,12 +1,14 @@
 #ifndef END_H
 #define END_H
 
-#include "con_lib.h"
-#include "board.h"
-#include "menu.h"
 #include <stdbool.h>
 #include <stdio.h>
 #include <stdlib.h>
+
+#include "con_lib.h"
+#include "board.h"
+#include "instance.h"
+#include "game_logic.h"
 
 bool hasWon(Board tempBoard);
 void loseScreen();
