+++ instance.h
@@ -1,5 +1,5 @@
-#ifndef MENU_H
-#define MENU_H
+#ifndef INSTANCE_H
+#define INSTANCE_H
 
 #include <stdlib.h>
 #include <stdio.h>
@@ -9,10 +9,14 @@
 #include "end.h"
 #include "con_lib.h"
 #include "save.h"
+#include "utility.h"
+#include "instance.h"
 
 bool wasPressed[128];
 int showMenu(char *menuTitle, char *menuOptions[], int menuSize, char *inputMsg);
 void menuInstance();
 void inGameMenu(Board *realBoard, int markedCount);
+void gameInstance(Board *realBoard);
+void showScoreboard();
 
 #endif
